Allowed to equip the shield
I went through and made it so the shield can be equipped as well. As of right now I just activate it then when I let go of the shield button it waits 5 seconds then disappears. If you press the shield button again it will stop that countdown and keep the shield.